Refactored collision, basically changed how it functioned.
Collision now uses a static class, CollisionTracker. This keeps track of
all collisiders and treats them according to their types, which is
defined in CollisionType. BoxCollision and CircularCollision are now
abstract classes, as is Collider. If you want to have a collider, you
make a nested class within the class using said collider, and extend
either BoxCollision or CircularCollision. See Player.java,
GenericEnemy.java, and Wall.java for examples.